Site now registers the user id/email and name in a database-table with a FK in Stockitem, as well as in the Vuex state. Users can now only view and manipulate their own list of stockitems. Still buggy, and thr 'delete type' button has been removed for further development. Also some minor style-changes and some code-cleanup.

+++ fridgein-api/Controllers/StockitemsController.cs
@@ -50,6 +50,31 @@ namespace fridgein_api.Controllers
             return Ok(grouped);
         }
 
+        // GET: api/stockitem/readuniqueonuser
+        [HttpGet]
+        [Route("readuniqueonuser/{email}")]
+        public async Task<ActionResult<IEnumerable<Stockitem>>> GetUniqueStockitemOnUser(string email)
+        {
+            User user = await _context.User.Where(u => u.Email == email).FirstAsync();
+
+            // Get list of all stockitems
+            ICollection<Stockitem> allItems = await _context.Stockitem
+                .Where(s => s.UserId == user.UserId)
+                .Include(s => s.Food)
+                .ToListAsync();
+
+            List<Stockitem> distinctItems = allItems
+                .GroupBy(s => new { s.FoodId, s.PurchaseDate, s.ExpirationDate })
+                .Select(g => g.First())
+                .ToList();
+
+            // Group by chosen fiels
+            var grouped = allItems.GroupBy(x => new { x.FoodId, x.PurchaseDate, x.ExpirationDate });
+
+
+            return Ok(grouped);
+        }
+
         // GET: api/stockitem/get/5
         [HttpGet("get/{id}")]
         public async Task<ActionResult<Stockitem>> GetStockitem(int id)
@@ -106,10 +131,10 @@ namespace fridgein_api.Controllers
         }
 
         // DELETE: api/stockitem/del/5
-        [HttpDelete("del/{id}")]
-        public async Task<ActionResult<Stockitem>> DeleteStockitem(int id)
+        [HttpDelete("del/{id}/{userid}")]
+        public async Task<ActionResult<Stockitem>> DeleteStockitem(int id, int userid)
         {
-            var stockitem = await _context.Stockitem.FindAsync(id);
+            var stockitem = await _context.Stockitem.Where(s => s.FoodId == id && s.UserId == userid).FirstAsync();
             if (stockitem == null)
             {
                 return NotFound();
@@ -123,10 +148,10 @@ namespace fridgein_api.Controllers
 
         // Delete on FoodId
         // DELETE: api/stockitem/delall/5
-        [HttpDelete("delall/{id}")]
-        public async Task<ActionResult<Stockitem>> DeleteAllStockitems(int id)
+        [HttpDelete("delall/{id}/{userId}")]
+        public async Task<ActionResult<Stockitem>> DeleteAllStockitems(int id, int userId)
         {
-             _context.Stockitem.RemoveRange(_context.Stockitem.Where(s => s.Food.FoodId == id));
+             _context.Stockitem.RemoveRange(_context.Stockitem.Where(s => s.Food.FoodId == id && s.UserId == userId));
             await _context.SaveChangesAsync();
             return Ok();
         }
